Refactor files and slight fixes
