+++ wordle-web/src/scripts/letter.ts
@@ -1,8 +1,8 @@
 export enum LetterStatus {
   NotGuessed = 0,
-  Correct,
-  Misplaced,
-  Wrong
+  Correct = 1,
+  Misplaced = 2,
+  Wrong = 3
 }
 
 export class Letter {
