+++ wordle-web/src/views/LeaderboardView.vue
@@ -1,7 +1,28 @@
 <template>
-  <BoardDisplay />
+  <h1>LeaderBoard</h1>
+  <div>
+    <v-carousel class="h-auto w-auto pb-13">
+      <v-carousel-item>
+        <v-card class="px-16 text-center">
+          <Leaderboard />
+        </v-card>
+      </v-carousel-item>
+      <v-carousel-item>
+        <v-card class="px-16 text-center">
+          <PersonalStats />
+        </v-card>
+      </v-carousel-item>
+    </v-carousel>
+  </div>
 </template>
 
 <script setup lang="ts">
-import BoardDisplay from '@/components/DisplayBoards/DisplayBoards.vue'
+import PersonalStats from '@/components/LeaderBoard/PersonalStatsCard.vue'
+import Leaderboard from '@/components/LeaderBoard/LeaderboardCard.vue'
 </script>
+
+<style>
+.v-card {
+  height: 100%;
+}
+</style>
